Move Predef to Dsl class
