+++ src/main/scala/uk/co/developmentanddinosaurs/gatling/lambda/Predef.scala
@@ -1,0 +1,3 @@
+package uk.co.developmentanddinosaurs.gatling.lambda
+
+object Predef extends LambdaDsl
